add a function for checking if drive exists
