+++ packages/msgraph/src/Utils.js
@@ -1,6 +1,25 @@
 import { fetch } from 'undici';
 import { composeNextState } from '@openfn/language-common';
 
+export function ensureDriveExistsOrFail(drive, fileType) {
+  if (!drive) {
+    const operation = fileType === 'folder' ? 'getFolder' : 'getFile';
+    const egJobCode = [
+      `Eg: Get a site drive then get "/Sample Data" ${fileType}`,
+      `getDrive({ id: "openfn.sharepoint.com", owner: "sites"})`,
+      `${operation}("/Sample Data/${fileType === 'folder' ? '' : 'test.csv'})`,
+    ].join('\n\t  ');
+
+    const errorString = [
+      `Message: Drive is not defined`,
+      `Quick Fix: Add getDrive() operation before ${operation}()`,
+      `${egJobCode}`,
+    ].join('\n\t∟ ');
+
+    throw new Error(errorString);
+  }
+}
+
 export function getUrl(resource, apiVersion) {
   if (isValidHttpUrl(resource)) return resource;
